Add files via upload
My code and what I had completed. I've included four screen shots. I could not get this code to work so I tried a multitude of bug fixes. The screen shots show I was able to get the code working correctly for certain parts at one time but not the rest and vis, versa. I think the directions were very difficult to follow considering it referenced a starting code I was not given.